MongoDB support for Event Sourcing apps

+++ source/content/castanha/MyProject.Infrastructure/MongoDataAccess/AccountRepository.cs
@@ -5,6 +5,8 @@
     using MongoDB.Driver;
     using System;
     using System.Threading.Tasks;
+    using MyProject.Domain.Accounts.Events;
+    using MyProject.Application;
 
     public class AccountRepository : IAccountReadOnlyRepository, IAccountWriteOnlyRepository
     {
@@ -15,13 +17,27 @@
             this.mongoContext = mongoContext;
         }
 
-        public async Task Add(Account account)
+        public async Task Add(OpenedDomainEvent domainEvent)
         {
+            Account account = new Account();
+            account.Apply(domainEvent);
+
             await mongoContext.Accounts.InsertOneAsync(account);
         }
 
-        public async Task Delete(Account account)
+        public async Task Delete(ClosedDomainEvent domainEvent)
         {
+            Account account = await mongoContext
+                .Accounts
+                .Find(e => e.Id == domainEvent.AggregateRootId)
+                .SingleOrDefaultAsync();
+
+            if (account == null)
+                throw new AccountNotFoundException($"The account {domainEvent.AggregateRootId} does not exists or is already closed.");
+
+            if (account.Version != domainEvent.Version)
+                throw new TransactionConflictException(account, domainEvent);
+
             await mongoContext.Accounts.DeleteOneAsync(e => e.Id == account.Id);
         }
 
@@ -33,8 +49,39 @@
                 .SingleOrDefaultAsync();
         }
 
-        public async Task Update(Account account, Transaction transaction)
+        public async Task Update(DepositedDomainEvent domainEvent)
         {
+            Account account = await mongoContext
+                .Accounts
+                .Find(e => e.Id == domainEvent.AggregateRootId)
+                .SingleOrDefaultAsync();
+
+            if (account == null)
+                throw new AccountNotFoundException($"The account {domainEvent.AggregateRootId} does not exists or is already closed.");
+
+            if (account.Version != domainEvent.Version)
+                throw new TransactionConflictException(account, domainEvent);
+
+            account.Apply(domainEvent);
+
+            await mongoContext.Accounts.ReplaceOneAsync(e => e.Id == account.Id, account);
+        }
+
+        public async Task Update(WithdrewDomainEvent domainEvent)
+        {
+            Account account = await mongoContext
+                .Accounts
+                .Find(e => e.Id == domainEvent.AggregateRootId)
+                .SingleOrDefaultAsync();
+
+            if (account == null)
+                throw new AccountNotFoundException($"The account {domainEvent.AggregateRootId} does not exists or is already closed.");
+
+            if (account.Version != domainEvent.Version)
+                throw new TransactionConflictException(account, domainEvent);
+
+            account.Apply(domainEvent);
+
             await mongoContext.Accounts.ReplaceOneAsync(e => e.Id == account.Id, account);
         }
     }
